+++ scripts/projects.ts
@@ -18,4 +18,5 @@ export const projects: string[] = [
 	'libs/webvtt',
 	'libs/drm',
 	'libs/cmaf-ham',
+	// 'libs/mse',
 ]